fix(checker): split global augmentation duplicate helper
