+++ crates/tsz-checker/src/types/type_checking/mod.rs
@@ -23,6 +23,7 @@ mod duplicate_identifiers;
 mod duplicate_identifiers_ambient_default;
 mod duplicate_identifiers_constructor;
 mod duplicate_identifiers_export_surface;
+mod duplicate_identifiers_global_augmentation;
 mod duplicate_identifiers_helpers;
 mod duplicate_identifiers_remote_lib;
 mod duplicate_index_signatures;
